machinedeployment: set revision on machinedeployment and patch via defer
